Reland "Making crashes uworker safe (#4102)" (#4137)
This reverts commit 8c2f3bcebd061dfdd40470ee2e2ddab6bf6637e1.

+++ src/clusterfuzz/_internal/bot/tasks/setup.py
@@ -782,11 +782,13 @@ def get_fuzzer_directory(fuzzer_name):
   return fuzzer_directory
 
 
-def archive_testcase_and_dependencies_in_gcs(resource_list, testcase_path):
-  """Archive testcase and its dependencies, and store in blobstore."""
+def archive_testcase_and_dependencies_in_gcs(resource_list, testcase_path: str,
+                                             upload_url: str):
+  """Archive testcase and its dependencies, and store in blobstore. Returns
+  whether it is archived, the absolute_filename, and the zip_filename."""
   if not os.path.exists(testcase_path):
-    logs.error('Unable to find testcase %s.' % testcase_path)
-    return None, None, None, None
+    logs.error(f'Unable to find testcase {testcase_path}.')
+    return None, None, None
 
   absolute_filename = testcase_path
   archived = False
@@ -804,7 +806,7 @@ def archive_testcase_and_dependencies_in_gcs(resource_list, testcase_path):
   # Filter out duplicates, directories, and files that do not exist.
   resource_list = utils.filter_file_list(resource_list)
 
-  logs.info('Testcase and related files :\n%s' % str(resource_list))
+  logs.info(f'Testcase and related files :\n{resource_list}')
 
   if len(resource_list) <= 1:
     # If this does not have any resources, just save the testcase.
@@ -812,8 +814,8 @@ def archive_testcase_and_dependencies_in_gcs(resource_list, testcase_path):
     try:
       file_handle = open(testcase_path, 'rb')
     except OSError:
-      logs.error('Unable to open testcase %s.' % testcase_path)
-      return None, None, None, None
+      logs.error(f'Unable to open testcase {testcase_path}.')
+      return None, None, None
   else:
     # If there are resources, create an archive.
 
@@ -830,7 +832,7 @@ def archive_testcase_and_dependencies_in_gcs(resource_list, testcase_path):
           break
 
     base_directory = os.path.sep.join(base_directory_list)
-    logs.info('Subresource common base directory: %s' % base_directory)
+    logs.info(f'Subresource common base directory: {base_directory}')
     if base_directory:
       # Common parent directory, archive sub-paths only.
       base_len = len(base_directory) + len(os.path.sep)
@@ -854,17 +856,20 @@ def archive_testcase_and_dependencies_in_gcs(resource_list, testcase_path):
     try:
       file_handle = open(zip_path, 'rb')
     except OSError:
-      logs.error('Unable to open testcase archive %s.' % zip_path)
-      return None, None, None, None
+      logs.error(f'Unable to open testcase archive {zip_path}.')
+      return None, None, None
 
     archived = True
     absolute_filename = testcase_path[base_len:]
 
-  fuzzed_key = blobs.write_blob(file_handle)
+  if not storage.upload_signed_url(file_handle, upload_url):
+    logs.error('Failed to upload testcase.')
+    return None, None, None
+
   file_handle.close()
 
   # Don't need the archive after writing testcase to blobstore.
   if zip_path:
     shell.remove_file(zip_path)
 
-  return fuzzed_key, archived, absolute_filename, zip_filename
+  return archived, absolute_filename, zip_filename
